Bridge: break up bridge_fn implementation by FFI/JNI/Node
Also, revert back to the simpler unzip3 crate instead of unzip-n.
