+++ rust/bridge/shared/macros/Cargo.toml
@@ -19,4 +19,4 @@ proc-macro2 = "1.0"
 quote = "1.0"
 syn = { version = "1.0" }
 syn-mid = "0.5"
-unzip-n = "0.1.2"
+unzip3 = "1.0"
